+++ src/Patches/NetPackageEntitySetSkillLevelServer.cs
@@ -14,7 +14,7 @@ namespace Amnesia.Patches
     {
         private static readonly ModLog<NetPackageEntitySetSkillLevelServer_ProcessPackage_Patches> _log = new ModLog<NetPackageEntitySetSkillLevelServer_ProcessPackage_Patches>();
 
-        public static void Postfix(int ___entityId, string ___skill, int ___level)
+        public static void Postfix(NetPackageEntitySetSkillLevelServer __instance, World _world, int ___entityId, string ___skill, int ___level)
         {
             try
             {
@@ -24,8 +24,21 @@ namespace Amnesia.Patches
                     _log.Error($"Unable to retrieve player record for entityId {___entityId}");
                     return;
                 }
-                playerStats.Changes.Add((___skill, ___level));
-                playerStats.Save();
+                if (!_world.Players.dict.TryGetValue(___entityId, out var player))
+                {
+                    _log.Error($"Unable to retrieve player for entityId {___entityId}");
+                    return;
+                }
+                _log.Trace($"Player {___entityId} {player.GetDebugName()} increased in level: {___skill} >> {___level}");
+                var progressionClass = player.Progression.GetProgressionValue(___skill).ProgressionClass;
+                if (progressionClass.IsAttribute || progressionClass.IsPerk) // don't track action skills or books
+                {
+                    playerStats.Changes.Add((___skill, ___level));
+                    playerStats.Save();
+                    // TODO: tell client to refresh server's skillPoints value now
+                    __instance.Sender.SendPackage(NetPackageManager.GetPackage<NetPackageConsoleCmdClient>().Setup("sm", true));
+                    __instance.Sender.SendPackage(NetPackageManager.GetPackage<NetPackageConsoleCmdClient>().Setup("sm", true));
+                }
             }
             catch (Exception e)
             {