+++ src/Utilities/PlayerHelper.cs
@@ -64,10 +64,13 @@ namespace Amnesia.Utilities
 
             if (Config.ForgetLevelsAndSkills)
             {
-                player.Progression.ResetProgression(Config.ForgetBooks);
-                player.Progression.Level = Config.LongTermMemoryLevel;
-                player.Progression.ExpToNextLevel = player.Progression.GetExpForNextLevel();
-                player.Progression.SkillPoints = Config.LongTermMemoryLevel - 1;
+                // TODO: come up with a new solution for books?
+                //  player.Progression.ResetProgression(Config.ForgetBooks);
+                //  player.Progression.Level = Config.LongTermMemoryLevel;
+                //  player.Progression.ExpToNextLevel = player.Progression.GetExpForNextLevel();
+                //  player.Progression.SkillPoints = Config.LongTermMemoryLevel - 1;
+
+                RewindLevelSkillProgression(player, (player.Progression.Level - Config.LongTermMemoryLevel) * Progression.SkillPointsPerLevel);
 
                 // Return all skill points rewarded from completed quest; should cover vanilla quest_BasicSurvival8, for example
                 if (!Config.ForgetIntroQuests)
@@ -183,6 +186,46 @@ namespace Amnesia.Utilities
             }));
             clientInfo.SendPackage(NetPackageManager.GetPackage<NetPackageEntityCollect>().Setup(entityId, clientInfo.entityId));
             _ = GameManager.Instance.World.RemoveEntity(entityId, EnumRemoveEntityReason.Despawned);
+        }
+
+        private static void RewindLevelSkillProgression(EntityPlayer player, int count)
+        {
+            _log.Trace($"RewindLevelSkillProgression({player.entityId}, {count})");
+            if (player.Progression.SkillPoints >= count)
+            {
+                _log.Trace($"player.Progression.SkillPoints >= {count})");
+                player.Progression.SkillPoints -= count;
+                player.Progression.Level -= count;
+                player.Progression.ExpToNextLevel = player.Progression.GetExpForNextLevel();
+                return;
+            }
+            _log.Trace($"player.Progression.SkillPoints < {count})");
+            var leftToReap = count - player.Progression.SkillPoints;
+
+            if (!PlayerRecord.Entries.TryGetValue(player.entityId, out var playerRecord))
+            {
+                _log.Error($"Could not rewind progression by {leftToReap} for player {player.entityId} {player.GetDebugName()}!");
+                return;
+            }
+            if (playerRecord.Changes.Count < leftToReap)
+            {
+                // TODO: handle this situation
+                // TODO: this should not hit in test... log output?
+                _log.Error($"count: {count} | leftToReap: {leftToReap} | skillPoints: {player.Progression.SkillPoints} | level: {player.Progression.Level}");
+                throw new NotImplementedException("TODO");
+            }
+
+            _log.Trace($"about to reverse {leftToReap} perks");
+            for (var i = playerRecord.Changes.Count - 1; i <= playerRecord.Changes.Count - 1 - leftToReap; i--)
+            {
+                _log.Trace($"reversing {playerRecord.Changes[i].Item1} level {playerRecord.Changes[i].Item2})");
+                player.Progression.GetProgressionValue(playerRecord.Changes[i].Item1).Level = playerRecord.Changes[i].Item2 - 1;
+            }
+            playerRecord.Changes.RemoveRange(playerRecord.Changes.Count - 1 - leftToReap, leftToReap);
+            player.Progression.SkillPoints = 0;
+            player.Progression.Level -= leftToReap;
+            player.Progression.ExpToNextLevel = player.Progression.GetExpForNextLevel();
+            playerRecord.Save();
         }
     }
 }